+++ R/population.R
@@ -10,6 +10,7 @@
 #' @examples
 #' ypr_population(k = 0.1, Linf = 90)
 ypr_population <- function(tmax = 20L, k = 0.15, Linf = 100, t0 = 0,
+                           k2 = k, Linf2 = Linf, L2 = 1000,
                            Wb = 3, Ls = 50, Sp = 100, es = 1, Sm = 0,
                            fb = 1,
                            tR = 1L, BH = 1L, Rk = 3, n = 0.2, nL = n,